requirement.txt
flask==2.2.3
langchain==0.15.0
langchain-community==0.0.1
langchain-huggingface==0.0.1
huggingface-hub==0.13.3
faiss-cpu==1.7.3
python-dotenv==0.21.1
requests==2.28.2
